requirements.txt
colorama==0.4.5
cycler==0.11.0
docx2pdf==0.1.8
et-xmlfile==1.1.0
fonttools==4.37.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.3
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pdf2image==1.16.0
Pillow==9.2.0
pyparsing==3.0.9
PyPDF2==2.10.6
python-dateutil==2.8.2
python-docx==0.8.11
pywin32==304
six==1.16.0
tqdm==4.64.1
XlsxWriter==3.0.3
